backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart